==> Code/02. Data Cleaning.ipynb <==
# import libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import statsmodels.api as sm
import seaborn as sns
from datetime import datetime

weather1 = pd.read_csv('../data/df_weather1.csv')
weather1.head()

weather1.shape

weather2 = pd.read_csv('../data/df_weather2.csv')
weather2.head()

weather2.shape

# concat weather1 and weather2
weather = pd.concat([weather1,weather2])
weather.head()

weather.shape

# remove unnamed column
weather.drop(columns=['Unnamed: 0'], inplace=True)
weather.head()

weather.info()

# check for null
weather.isnull().sum()

# drop null as it is less than 5% of whole dataset
weather.dropna(inplace=True)

weather.isnull().sum()

weather.apply(lambda col: col.str.contains('-'), axis=1).sum()

# Replace '-' with 0
def float_checker(row):
    try:
        res = float(row)
    except:
        res = 0
    return res

c = ['Daily Rainfall Total (mm)',\
             'Highest 30-min Rainfall (mm)',\
             'Highest 60-min Rainfall (mm)',\
             'Highest 120-min Rainfall (mm)',\
             'Mean Temperature (°C)',\
             'Maximum Temperature (°C)',\
             'Minimum Temperature (°C)',\
             'Mean Wind Speed (km/h)',\
             'Max Wind Speed (km/h)']

for col in c:
    weather[col] = weather[col].map(float_checker)

weather.info()

weather['Date'] = weather['Date']+' '+weather['year'].astype(str)

weather["Date"]= pd.to_datetime(weather["Date"])

# change column name by replacing ' (mm)' with ''
weather.columns = weather.columns.str.replace(' (mm)', '', regex=False)

# change column name by replacing ' (°C)' with ''
weather.columns = weather.columns.str.replace(' (°C)', '', regex=False)

# change column name by replacing ' (km/h)' with ''
weather.columns = weather.columns.str.replace(' (km/h)', '', regex=False)

# change column name by replacing '-' with '_'
weather.columns = weather.columns.str.replace('-', '_', regex=False)

# change column name to lowercase and replace ' ' with '_'
weather.columns = weather.columns.str.replace(' ', '_', regex=False).str.lower()

weather.head()

weather.info()

**Create 14 rolling average for each region**

region_names = dict()
for k, v in weather.groupby('region'):
    region_names[k] = v

region_names['Admiralty']

def rolling_mean(df):
    df['daily_rainfall_total_14'] = df['daily_rainfall_total'].rolling(14).mean()
    df['highest_30_min_rainfall_14'] = df['highest_30_min_rainfall'].rolling(14).mean()
    df['highest_60_min_rainfall_14'] = df['highest_60_min_rainfall'].rolling(14).mean()
    df['highest_120_min_rainfall_14'] = df['highest_120_min_rainfall'].rolling(14).mean()
    df['mean_temperature_14'] = df['mean_temperature'].rolling(14).mean()
    df['maximum_temperature_14'] = df['maximum_temperature'].rolling(14).mean()
    df['minimum_temperature_14'] = df['minimum_temperature'].rolling(14).mean()
    df['mean_wind_speed_14'] = df['mean_wind_speed'].rolling(14).mean()
    df['max_wind_speed_14'] = df['max_wind_speed'].rolling(14).mean()
    return df

for r in region_names.values():
    r.sort_values('date', ascending = True, inplace=True)
    rolling_mean(r)

region_names['Admiralty']

region_list = []
for i in region_names.keys():
    temp_df = region_names[i]
    region_list.append(temp_df)

weather = pd.concat(region_list)

weather.head()

weather.shape

# save data
weather.to_csv('../data/weather.csv', index=False)
