==> Code/03. dengue_cases_cleaning.ipynb <==
import pandas as pd
import numpy as np
import glob

#### Data import and cleaning

Data Source: https://outbreak.sgcharts.com/data

cases_df = pd.DataFrame()
shape = 0
for file in glob.glob('../Data/sgcharts/csv/*'):
    df = pd.read_csv(file,header = None)
    shape += df.shape[0]
    cases_df = cases_df.append(df)

print(shape)
cases_df.head()

cases_df2 = pd.DataFrame()
shape = 0
for file in glob.glob('../Data/sgcharts/incorrect_latitude_longitude/*.csv'):
    df = pd.read_csv(file,header = None)
    shape += df.shape[0]
    cases_df2 = cases_df2.append(df)

print(shape)
cases_df2.head()

cases_df.shape

cases_df.columns = ['no_of_cases','address','latitude','longitude','cluster_no','recent_cases','total_cases','date','month']

cases_df.reset_index(drop = True, inplace = True)

cases_df.tail()

cases_df['date'] = pd.to_datetime(cases_df['date'], format='%y%m%d')

cases_df['recent_cases'].value_counts()

# pip install geopy

# Remove duplicates so that the reverse geocoding runs faster
coords_to_map = cases_df.drop_duplicates(['address'])[['address','latitude','longitude']]
coords_to_map.shape

import geopy

def get_zipcode(df, geolocator, lat_field, lon_field):
    try:
        location = geolocator.reverse((df[lat_field], df[lon_field]))
        return location.raw['address']['postcode']
    except: 
        return 0

def get_suburb(df, geolocator, lat_field, lon_field):
    try:
        location = geolocator.reverse((df[lat_field], df[lon_field]))
        return location.raw['address']['suburb']
    except: 
        return "-"


geolocator = geopy.Nominatim(user_agent='1234')

coords_to_map['postal_code'] = coords_to_map.apply(get_zipcode, axis=1, geolocator=geolocator, lat_field='latitude', lon_field='longitude')
coords_to_map['suburb'] = coords_to_map.apply(get_suburb, axis=1, geolocator=geolocator, lat_field='latitude', lon_field='longitude')

coords_to_map

cases_df2 = cases_df.merge(coords_to_map[['address','postal_code','suburb']], how = 'left', on = 'address')

Map the remaining postal codes to their respective area if the reverse geocoder did not manage to return an area
([*source*](https://www.ura.gov.sg/realEstateIIWeb/resources/misc/list_of_postal_districts.htm))

mapping = pd.read_csv('../Data/postal_mapping.csv')

mapping.head()

def mapping_postal(row):
    if row['suburb'] == "-":
        postal = row['postal_code'][:2]
        return mapping.loc[mapping['Postal Sector'].str.contains(postal),'General Location'].reset_index()['General Location'][0]
    else:
        return row['suburb']

cases_df2['suburb'] = cases_df2.apply(mapping_postal,axis=1)

# Save a WIP ver just in case, as the reverse geocoding takes hours to run
cases_df2.to_csv('../Data/dengue_cases_region_wip.csv', index = False)

# Open the WIP ver to skip the above steps
cases_df2 = pd.read_csv('../Data/dengue_cases_region_wip.csv')

cases_region = cases_df2['suburb'].unique()
cases_region

# Read in the weather data's region to clean up the suburb
weather = pd.read_csv('../Data/weather.csv')
region_list = weather['region'].unique()
region_list

print('These areas should be renamed: ')
for i in cases_region:
    if i not in region_list:     
        print(i)

cases_df3 = cases_df2.replace({
    'Bishan': 'Ang Mo Kio',
    'Geylang': 'Tai Seng',
    'Tampines': 'Simei',
    'Bedok': 'Chai Chee',
    'Jurong West': 'Jurong (West)',
    'Holland Village': 'Buona Vista',
    'Jurong East': 'Jurong (East)',
    'Woodlands': 'Admiralty',
    'Choa Chu Kang': 'Choa Chu Kang (Central)',
    'Seng Kang': 'Buangkok',
    'Hougang': 'Buangkok',
    'Bukit Merah': 'Kampong Bahru',
    'Tanglin': 'Botanic Garden',
    'Queenstown, Tiong Bahru': 'Queenstown',
    'Pasir Ris': 'Pasir Ris (Central)',
    'Bukit Batok': 'Bukit Panjang',
    'Rochor': 'Dhoby Ghaut',
    'Jurong': 'Jurong (East)',
    'Kallang': 'Whampoa',
    'Joo Seng': 'Tai Seng',
    'Pasir Panjang, Hong Leong Garden, Clementi New Town': 'Pasir Panjang',
    'Downtown Core': 'Marina Barrage',
    'Novena': 'Newton',
    'Serangoon Gardens': 'Serangoon North',
    'Upper Bukit Timah, Clementi Park, Ulu Pandan': 'Ulu Pandan',
    'Hillview': 'Bukit Panjang',
    'Changi Business Park': 'Changi',
    'Kaki Bukit': 'Chai Chee',
    'one-north': 'Buona Vista',
    'Kallang Way': 'Whampoa',
    'Sungei Kadut': 'Kranji Reservoir',
    'Singapore River': 'Dhoby Ghaut',
    'Outram': 'Tanjong Pagar',
    'Potong Pasir': 'Toa Payoh',
    'Western Islands': 'Jurong Island',
    'Western Water Catchment': 'Tengah',
    'Central Water Catchment': 'Macritchie Reservoir',
    'Katong': 'Marine Parade',
    'Pioneer': 'Jurong (West)',
    'Orchard': 'Somerset (Road)',
    'River Valley': 'Somerset (Road)',
    'Alexandra': 'Queenstown',
    'Museum': 'Dhoby Ghaut'
})

cases_df3['suburb'].value_counts()

cases_df3['suburb'].nunique()

cases_df3.head()

# Drop unncessary columns
cases_df3.drop(['month'],axis=1,inplace=True)
cases_df3.rename({'suburb':'region'},axis=1,inplace=True)
cases_df3.head()

cases_df3['date'] = pd.to_datetime(cases_df3['date'])

cases_df3.info()

cases_df3.to_csv('../Data/dengue_cases_region.csv', index = False)

cases_df3 = pd.read_csv('../Data/dengue_cases_region.csv')

#### Data Dictionary

|Feature|Type|Dataset|Description|
|:---|:---|:---|:---|
|no_of_cases|integer|dengue cluster (NEA)|Number of reported dengue cases at this location| 
|address|object|dengue cluster (NEA)|Street address where dengue cases are reported (down to the apartment block level)| 
|latitude|float|dengue cluster (NEA)|Latitude of the street address| 
|longitude|float|dengue cluster (NEA)|Longitude of the street address| 
|cluster_no|int|dengue cluster (NEA)|Every dengue cluster is labelled with a serial number. However, this serial number cannot be used as an unique identifier because (1) the serial number is reused in other snapshots and (2) the serial number will change throughout the cluster's lifetime|
|recent_cases|integer|dengue cluster (NEA)|Number of dengue cases with onset in the last 2 weeks| 
|total_cases|integer|dengue cluster (NEA)|Total number of dengue cases reported in this cluster| 
|date|datetime|dengue cluster (NEA)|Date in YYYY-MM-DD format| 
|postal_code|integer|dengue cluster (NEA)|Postal code of the stress address| 
|region|object|dengue cluster (NEA) & postal mapping (URA)|Region of the street address| 

res_df = cases_df3.drop_duplicates(subset=['cluster_no', 'region','date']).groupby(['region','date'])['recent_cases'].sum().reset_index()
res_df.head()

NEA published the count of recent cases per cluster only from December 2013 onwards. For prior data, this field is substituted with a placeholder value of -1.

Hence, remove data with cases < 0 

res_df[res_df['recent_cases'] < 0]['date'].unique()

res_df2 = res_df[res_df['recent_cases'] > 0]

Change the dates to the nearest Sunday so that we are able to match with the weekly Google trends data

res_df2['date'] = pd.to_datetime(res_df2['date'])

import datetime
def datetime_to_next_sunday(row):
    if row.weekday() <= 2: # if date is mon/tue/wed, turn date into previous sun
        return row - datetime.timedelta(days=row.weekday()+1)
    else: # if date is thu/fri/sat, turn date into next sun
        return row + datetime.timedelta(days=6-row.weekday())

res_df2['date_sunday'] = res_df2[['date']].applymap(datetime_to_next_sunday)

res_df2.head()

Check to see if all dates are Sunday

res_df2['date_sunday'].apply(lambda x: x.weekday()).unique()

res_df3 = res_df2.drop('date',axis=1)
res_df3.rename({'date_sunday':'date'},axis=1,inplace=True)

res_df3.tail()

res_df3.to_csv('../Data/dengue_cases_region_final.csv', index = False)
