==> Code/04. google_trend_cleaning.ipynb <==
# Import libraries
import pandas as pd

# Read Google Trend data
aedes = pd.read_csv('../Data/Trends/trend_aedes.csv')
dengue = pd.read_csv('../Data/Trends/trend_dengue.csv')
fever = pd.read_csv('../Data/Trends/trend_fever.csv')
headache = pd.read_csv('../Data/Trends/trend_headache.csv')
nosebleed = pd.read_csv('../Data/Trends/trend_nosebleed.csv')
vomit = pd.read_csv('../Data/Trends/trend_vomit.csv')

# Function to rename the date column and set it to index
def indexing (df):
    df.rename(columns={'Category: All categories': 'date'}, inplace=True)
    df.set_index('date', inplace=True)
    return df

# List of Google Trend data
trends = [aedes,dengue,fever,headache,nosebleed,vomit]

# Rename column and set date to index for all the Google Trend data
for trend in trends:
    trend = indexing(trend)
    print (trend.shape)

# Concatenate the data into a single dataframe
trend = pd.concat([aedes,dengue,fever,headache,nosebleed,vomit], axis=1)

# Display the concatenated dataframe
trend.head()

# Rename columns and remove the first 2 empty rows
trend.columns = ['aedes','dengue','fever','headache','nosebleed','vomit']
trend = trend.iloc[2:,:]

# reset index and change to datetime format.
trend = trend.reset_index()
trend['date'] = pd.to_datetime(trend['date'])
trend.head()

# Export to google_trends.csv
trend.to_csv('../Data/google_trends.csv')
