==> Code/05. merge_data.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 100)

# Read dengue csv file
dengue = pd.read_csv('../Data/dengue_cases_region_final.csv')
dengue.shape

# Display dengue dataframe
dengue.head()

# Convert date column to datetime format and set as index.
dengue.date = pd.to_datetime(dengue.date)
dengue.set_index('date', inplace=True)

# Check number of unique regions
regions = list(dengue.region.unique())
len(regions)

# Function to split the dengue dataframe based on region and sum up cases based on region
def split_df (dataframe, region):
    df = pd.DataFrame(dataframe[dataframe['region']==region]['recent_cases'])
    df = df.groupby('date').sum()
    df['region'] = region
    return df

# Perform the function and concatenate into a single dataframe
dengues = pd.DataFrame()
for r in regions:
    d = split_df(dengue, r)
    dengues = pd.concat([dengues, d], axis=0)

# Reset the index in order to merge with other dataframe easily
dengues=dengues.reset_index()

# Read both Google trends and Weather report
trends = pd.read_csv('../Data/google_trends.csv')
weather = pd.read_csv('../Data/weather.csv')

# Display Google trends data
trends.head()

trends.drop(columns='Unnamed: 0', inplace=True)

# Display weather data
weather.head()

# Convert date columns into datetime format
trends['date'] = pd.to_datetime(trends['date'])
weather['date'] = pd.to_datetime(weather['date'])

# Merge all data into a single dataframe
new_df = pd.merge(weather, trends, how ='left',on=['date'])
new_df = pd.merge(new_df, dengues, how ='left', on=['date','region'])

# Display merged data
new_df.tail()

# Export to a csv file for EDA and modelling.
new_df.to_csv('../Data/merged_data.csv')
