==> Code/09. cost_benefit_analysis.ipynb <==
### Fumigation Cost
---

Taking sample from four town's councils' tender for provision of mosquito, it seems like a three year contract cost around one million SGD.<br><br>
In Singapore, each town council (GRC) consist on average 5 estates.<br><br>
We assume that the vendor is expected to perform the fumigation twice a month, which tallies with NEA's logic of collecting fornightly case number.<br><br>
The estimated cost of one fumigation session for one estate will the work out to about SGD2,778.

|                                                                                    	|                                     	|
|------------------------------------------------------------------------------------	|-------------------------------------	|
|              Each Town Council 3-years Mosquito Contract Average Cost              	|            SGD 1,000,000            	|
| Each town council (GRC) consists on average 5 estates. Hence, cost for each estate 	|   SGD 1,000,000 / 5 = SGD 200,000   	|
|      Assume fumigation twice a month (fortnightly) Cost per fumigation session     	| SGD 200,000 / 36 / 2 = SGD 2,777,78 	|

### Treatment Cost
---

In a notice published by MOH, it stated that before Government subsidy, a dengue blood test could cost about SGD 70 - SGD 80 at a General Practitioner (GP) clinic, and about SGD 50 - SGD 60 at a polyclinic.<br>

|       Type of Clinic      	| Dengue Blood Test Cost (SGD) 	|
|:-------------------------:	|:----------------------:	|
| General Practitioner (GP) 	|        70 - 80       	|
|         Polyclinic        	|       50 - 60       	|

Given that the queue at polyclinics are usually very long, we assume that individuals go to clinics once they have dengue symptoms.<br>
We will take the median cost, that SGD 75 for adengue blood test.<br><br>

In addition, a regular consultation in clinics cost about SGD 20 to SGD 55.<br>
Taking the median again, we assume the cost of consultation to be SGD 37.50.<br><br>

The total cost for dengue treatment will add up to SGD 112.50.<br>

| Treatment Service 	| Cost (SGD) 	|
|:-----------------:	|:----------:	|
| Blood Test        	| 75         	|
| Consultation      	| 37.5       	|
|     **Total**     	|  **112.5** 	|

### Breakeven Number of Cases
---

The breakeven number of dengue for fumigation would therefore be:<br>
SGD2,778 / SGD 112.50 = 25 cases

|             Items             	| Cost (SGD) 	|
|:-----------------------------:	|:----------:	|
| Fumigation (every two weeks)  	| 2,778      	|
| Treatment                     	| 112.5      	|
| **Breakeven number of cases** 	|   **25**   	|


### Top 10 Regions where more than 25 cases per fortnight were recorded
---

|    	|               	|                   	|
|----	|---------------	|-------------------	|
| #  	|  region       	|  no_of_fortnights 	|
|  1 	| Chai Chee     	|         56        	|
|  2 	| Ang Mo Kio    	|         50        	|
|  3 	| Buangkok      	|         35        	|
|  4 	| Serangoon     	|         35        	|
|  5 	| Tai Seng      	|         35        	|
|  6 	| Bukit Panjang 	|         30        	|
|  7 	| Jurong (West) 	|         20        	|
|  8 	| Toa Payoh     	|         20        	|
|  9 	| Admiralty     	|         17        	|
| 10 	| Simei         	|         16        	|

### Top 10 Regions with Highest Average Case per Fortnight
---

|    	|                 	|              	|           	|
|----	|-----------------	|--------------	|-----------	|
| #  	|  region         	|  total_cases 	|  avg_case 	|
|  1 	| Ang Mo Kio      	|       3345.0 	|     35.59 	|
|  2 	| Chai Chee       	|       4793.0 	|     34.24 	|
|  3 	| Serangoon       	|       2732.0 	|     31.77 	|
|  4 	| Tai Seng        	|       2860.0 	|     28.04 	|
|  5 	| Toa Payoh       	|       1804.0 	|     23.43 	|
|  6 	| Bukit Panjang   	|       2131.0 	|      20.1 	|
|  7 	| Buangkok        	|       2764.0 	|     18.07 	|
|  8 	| Admiralty       	|       1837.0 	|     16.85 	|
|  9 	| Somerset (Road) 	|        176.0 	|      16.0 	|
| 10 	| Marine Parade   	|       1058.0 	|     15.79 	|

*Source:*<br>
1. [Marine Parade Town Council Tender](https://www.mptc.org.sg/TenderResult)
2. [Bishan-Toa Payoh Town Council Tender](https://www.btptc.org.sg/Resources/Documents/1702fdb4f2674c26905c104ae88ce66c.pdf)
3. [Choa Chu Kang Town Council Tender](https://www.ccktc.org.sg/wp-content/uploads/2022/04/CCKTC-TC-22-6911.pdf)
4. [Sembawang Town Council Tender](https://sbtc.org.sg/bo/web/upload/images/2APTR-BM-22-012-SB-R1672285420.pdf)
5. [Dengue Fever: How Much Treatment And Insurance Costs In Singapore](https://www.womensweekly.com.sg/beauty-and-health/dengue-fever-symptoms-prevention-treatment-costs/)
6. [Difference In Cost of Testing For Dengue Fever At Chas Clinics, Polyclinics And Private Clinics](https://www.moh.gov.sg/news-highlights/details/difference-in-cost-of-testing-for-dengue-fever-at-chas-clinics-polyclinics-and-private-clinics/)
